refactor: 불필요한 응답 필드 제거
