+++ src/main/kotlin/com/petqua/application/banner/dto/BannerDtos.kt
@@ -1,14 +1,11 @@
-package com.petqua.application.banner
+package com.petqua.application.banner.dto
 
 import com.petqua.domain.banner.Banner
-import java.time.LocalDateTime
 
 data class BannerResponse(
     val id: Long,
     val imageUrl: String,
     val linkUrl: String,
-    val createAt: LocalDateTime,
-    val updateAt: LocalDateTime,
 ) {
     companion object {
         fun from(banner: Banner): BannerResponse {
@@ -16,8 +13,6 @@ data class BannerResponse(
                 id = banner.id,
                 imageUrl = banner.imageUrl,
                 linkUrl = banner.linkUrl,
-                createAt = banner.createdAt,
-                updateAt = banner.updatedAt,
             )
         }
     }
