+++ src/main/kotlin/com/petqua/presentation/banner/BannerController.kt
@@ -1,7 +1,7 @@
 package com.petqua.presentation.banner
 
-import com.petqua.application.banner.BannerResponse
 import com.petqua.application.banner.BannerService
+import com.petqua.application.banner.dto.BannerResponse
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
